update additional endpoints
